+++ frontend/catalyst/jax_tracer.py
@@ -647,12 +647,9 @@ def trace_quantum_operations(
     for op in ops:
         qrp2 = None
         if isinstance(op, HybridOp):
-            kwargs = (
-                {"postselect_mode": mcm_config.postselect_mode}
-                if isinstance(op, catalyst.api_extensions.quantum_operators.MidCircuitMeasure)
-                else {}
-            )
-            qrp2 = op.trace_quantum(ctx, device, trace, qrp, **kwargs)
+            qrp2 = op.trace_quantum(ctx, device, trace, qrp)
+        elif isinstance(op, catalyst.api_extensions.quantum_operators.MidCircuitMeasure):
+            qrp2 = op.trace_quantum(ctx, trace, qrp, postselect_mode=mcm_config.postselect_mode)
         elif isinstance(op, MeasurementProcess):
             qrp2 = qrp
         else:
@@ -1102,7 +1099,9 @@ def trace_quantum_function(
                 return_values = tree_unflatten(out_tree_promise(), return_values_flat)
 
             def is_leaf(obj):
-                return isinstance(obj, qml.measurements.MeasurementProcess)
+                return isinstance(obj, qml.measurements.MeasurementProcess) and not isinstance(
+                    obj, qml.measurements.MidMeasureMP
+                )
 
             # 2. Create a new tree that has measurements as leaves
             return_values_flat, return_values_tree = jax.tree_util.tree_flatten(