+++ doc/changelog.md
@@ -7,7 +7,7 @@
 
 * The `dynamic_one_shot` transform uses a single auxiliary tape which is repeatedly simulated `n_shots` times to simulate hardware-like results.
   The loop over shots is executed with `catalyst.vmap`.
-  [(#5617)](https://github.com/PennyLaneAI/pennylane/pull/5617)
+  [(#773)](https://github.com/PennyLaneAI/catalyst/pull/773)
 
 * The Catalyst frontend now supports Python logging through PennyLane's `qml.logging` module.
   [(#660)](https://github.com/PennyLaneAI/catalyst/pull/660)
@@ -319,6 +319,10 @@
   [(#820)](https://github.com/PennyLaneAI/catalyst/pull/820)
 
 <h3>Internal changes</h3>
+
+* `MidCiruitMeasure` now inherits `MidMeasureMP` instead of `HybridOp`, allowing much better integration with PennyLane for
+  workflows with mid-circuit measurements.
+  [(#832)](https://github.com/PennyLaneAI/catalyst/pull/832)
 
 * Catalyst uses the `collapse` method of Lightning simulators in `Measure` to select a state vector branch and normalize.
   [(#801)](https://github.com/PennyLaneAI/catalyst/pull/801)